+++ docs/html/search/all_6.js
@@ -1,6 +1,19 @@
 var searchData=
 [
-  ['frequency_0',['Frequency',['../group__freq.html',1,'']]],
-  ['frequency_2eh_1',['frequency.h',['../frequency_8h.html',1,'']]],
-  ['functions_2',['Use STC15 HAL functions',['../index.html#autotoc_md4',1,'']]]
+  ['enable_5fadc_5finterrupt_0',['enable_adc_interrupt',['../group__interrupts.html#gaadf84a2181fd871ee2e29848480756ae',1,'interrupt.h']]],
+  ['enable_5fint0_5finterrupt_1',['enable_int0_interrupt',['../group__interrupts.html#ga9f0976519e8c6f8102dce44f7ad2aa6a',1,'interrupt.h']]],
+  ['enable_5fint1_5finterrupt_2',['enable_int1_interrupt',['../group__interrupts.html#ga4fb0d9228dcf2749cd098092edaa1308',1,'interrupt.h']]],
+  ['enable_5fint2_5finterrupt_3',['enable_int2_interrupt',['../group__interrupts.html#ga590f0de8191271977e1f79d9ab9576ab',1,'interrupt.h']]],
+  ['enable_5fint3_5finterrupt_4',['enable_int3_interrupt',['../group__interrupts.html#ga10ccfe3b81e2f0f8f8a9014a6f6cb382',1,'interrupt.h']]],
+  ['enable_5fint4_5finterrupt_5',['enable_int4_interrupt',['../group__interrupts.html#ga0e5e4a9cb47201c8f5be1e0918d29ac4',1,'interrupt.h']]],
+  ['enable_5flow_5fvoltage_5finterrupt_6',['enable_low_voltage_interrupt',['../group__interrupts.html#gada95baa6c6b2484eb8cf3c01889071fc',1,'interrupt.h']]],
+  ['enable_5fmaster_5fclock_5foutput_5fdiv1_7',['enable_master_clock_output_div1',['../group__freq.html#gab0cf04749c06459a905501a82de37ec5',1,'frequency.h']]],
+  ['enable_5fmaster_5fclock_5foutput_5fdiv2_8',['enable_master_clock_output_div2',['../group__freq.html#gab8cf6beb70bdf87d5ecb400ed00bb797',1,'frequency.h']]],
+  ['enable_5fmaster_5fclock_5foutput_5fdiv4_9',['enable_master_clock_output_div4',['../group__freq.html#ga6fa1489df98df6cf04b7ef07168267a2',1,'frequency.h']]],
+  ['enable_5fmcu_5finterrupts_10',['enable_mcu_interrupts',['../group__interrupts.html#ga45385330048db93cc8070cac1819be5c',1,'interrupt.h']]],
+  ['enable_5fspi_5finterrupt_11',['enable_spi_interrupt',['../group__interrupts.html#ga1031a01617faf8bd48f91d9eac505792',1,'interrupt.h']]],
+  ['enable_5ftimer0_5finterrupt_12',['enable_timer0_interrupt',['../group__interrupts.html#ga3ce3cdb9bf4963e75a0664964fb83c3a',1,'interrupt.h']]],
+  ['enable_5ftimer2_5finterrupt_13',['enable_timer2_interrupt',['../group__interrupts.html#ga6ba6fe72d49d266c76ca27b31bcb0395',1,'interrupt.h']]],
+  ['enable_5fuart1_5finterrupt_14',['enable_uart1_interrupt',['../group__interrupts.html#ga650e12d0892198b3a63c1a8c382b9c12',1,'interrupt.h']]],
+  ['external_5finterrupt_5ftrigger_5ft_15',['external_interrupt_trigger_t',['../group__interrupts.html#ga97fc91f85abe6139a7f24fbdff35c7e7',1,'interrupt.h']]]
 ];