+++ docs/html/search/enums_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['interrupt_5fpriority_5ft_0',['interrupt_priority_t',['../group__interrupts.html#ga9ebbbf91c4e6fa77cced0ac0daefd964',1,'interrupt.h']]]
+  ['external_5finterrupt_5ftrigger_5ft_0',['external_interrupt_trigger_t',['../group__interrupts.html#ga97fc91f85abe6139a7f24fbdff35c7e7',1,'interrupt.h']]]
 ];