+++ docs/html/search/searchdata.js
@@ -1,9 +1,9 @@
 var indexSectionsWithContent =
 {
-  0: "0_acdefghilmnoprstuvwy",
+  0: "0_abcdefghilmnoprstuvwy",
   1: "cdfiprstuw",
   2: "_cgmpu",
-  3: "eimt",
+  3: "beimt",
   4: "hlnoprt",
   5: "clm",
   6: "0acdfimprtuw",